Fix security issue
The problem is that registered user with command like the following:

curl ­-X PUT -­d 'post[user_id]=4' -­u 'gates:123' -­i \
http://api.example.com:3000/posts/1

can try to change creator of his post or comment, to
somebody else. This is not allowed.

The solution is 404 (Not Found) response instead of 'HTTP/1.1 500
Internal Server Error' in this case.